KimK2.5_From_Agent_Swarms_to_Production_Pipelines/requirements.txt
openai>=1.0.0
python-dotenv
opencv-python-headless
pillow
requests
